keygen/requirements.txt
altgraph==0.17.4
install==1.3.5
packaging==24.0
pefile==2023.2.7
py-machineid==0.5.1
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.4
PyJWT==2.8.0
PyQt6==6.6.1
PyQt6-Qt6==6.6.3
PyQt6-sip==13.6.0
pywin32-ctypes==0.2.2
setuptools==69.5.1
winregistry==1.1.1
